LPD-90205 Move the Data Masks table columns to a Java FDSView
