+++ modules/apps/mcp/mcp-server-web/src/main/resources/META-INF/resources/js/props_transformer/utils/maskEditURL.ts
